Move tokenizer endpoint and functions to separate file
